+++ tests/testthat/test-vimp.R
@@ -35,12 +35,29 @@ test_that("VIMP ensures even number of splits", {
     )
 })
 
+test_that("linear VIMP works with weights", {
+    d <- make_splits(d, uid, .num_splits = 4)
+    d$w <- rexp(n, 1 / 0.9) + 0.1
+    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new()))
+    expect_error(result <- calculate_linear_vimp(
+        d, w, y, cov1, cov2, cov3,
+        .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp, .Model_cfg = SLEnsemble_cfg$new()
+    ), NA)
+
+    expect_gt(result$estimate[1], result$estimate[2])
+    expect_gt(result$estimate[3], result$estimate[2])
+
+    expect_gt(result$estimate[1] / result$std_error[1], 2)
+    expect_gt(result$estimate[3] / result$std_error[3], 2)
+    expect_lt(result$estimate[2] / result$std_error[2], 2)
+})
+
 test_that("VIMP works with weights", {
     d <- make_splits(d, uid, .num_splits = 4)
     d$w <- rexp(n, 1 / 0.9) + 0.1
     attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new()))
     expect_error(result <- calculate_vimp(
-        d, w, y, cov1, cov2, cov3, 
+        d, w, y, cov1, cov2, cov3,
         .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp, .Model_cfg = SLEnsemble_cfg$new()
     ), NA)